AvinaPlan/requirements.txt
Django
djangorestframework
drf_yasg
psycopg2
Pillow
persian
Pyjwt
requests
python-dotenv
django-otp
django-otp-yubikey